Rename navigation folder and modify jest.config.js
