+++ App.tsx
@@ -1,7 +1,7 @@
 import 'react-native-gesture-handler';
 import { StatusBar } from "expo-status-bar";
 import { StyleSheet, Text, View } from "react-native";
-import Navigation from "./navigatons";
+import Navigation from "./navigations";
 
 export default function App() {
   return (
