+++ jest.config.js
@@ -10,6 +10,7 @@ module.exports = {
       '!**/node_modules/**',
       '!**/babel.config.js',
       '!**/jest.config.js',
+      '!**/navigations/**',
       '!**/utils/**',
       '!**/helper.ts',
       '!**/App.tsx',
